Upgrade forked @opentui/core from 0.1.88 to 0.2.1
Replace the fork base with upstream 0.2.1 and re-apply all Fermi-specific
customizations on top:

- Scroll gesture caching (48ms hitTest reuse window in renderer)
- Mousedown order (dispatchMouseEvent before startSelection, respecting
  preventDefault)
- ScrollBox batched scroll (pendingWheelDelta + onLifecyclePass,
  canConsumeScrollDirection boundary check, autoHideScrollbars,
  stopPropagation/preventDefault)
- EditBufferRenderable batched scroll + cursor scissor clipping
- Box dividerRatio/dividerTitle/titleColor properties
- Buffer JS-side scissor stack for cursor visibility queries
- Markdown getInterBlockMargin forced to 0, cellPaddingX/Y split
- Textarea contentChanged() yoga + viewport clamp overrides
- zig.ts local dylib candidate resolution before npm fallback
- Diagnostic system (lib/diagnostic.ts + renderer instrumentation)

Deps synced to upstream requirements: diff ^9, marked ^17,
strip-ansi ^7.1, web-tree-sitter 0.25.10.

RGBA migration: upstream 0.2.1 natively supports indexed colors via
RGBA.fromIndex() + intent/slot API (Uint16Array packing), replacing our
Float32Array sentinel encoding. Callsite updated.

Tree-sitter fix: dynamic import of web-tree-sitter with pre-set
globalThis.Module.wasmBinary to bypass Bun's __dirname doubling in
hoisted .bun/ packages.

+++ opentui-src/forked/core/testing/test-recorder.test.ts
@@ -1,4 +1,3 @@
-// @ts-nocheck
 import { describe, test, expect, beforeEach, afterEach } from "bun:test"
 import { createTestRenderer, type TestRenderer } from "./test-renderer.js"
 import { TestRecorder } from "./test-recorder.js"
@@ -293,7 +292,7 @@ describe("TestRecorder", () => {
     const frames = recorderWithFg.recordedFrames
     expect(frames.length).toBe(1)
     expect(frames[0].buffers).toBeDefined()
-    expect(frames[0].buffers?.fg).toBeInstanceOf(Float32Array)
+    expect(frames[0].buffers?.fg).toBeInstanceOf(Uint16Array)
     expect(frames[0].buffers?.bg).toBeUndefined()
     expect(frames[0].buffers?.attributes).toBeUndefined()
 
@@ -311,7 +310,7 @@ describe("TestRecorder", () => {
     const frames = recorderWithBg.recordedFrames
     expect(frames.length).toBe(1)
     expect(frames[0].buffers).toBeDefined()
-    expect(frames[0].buffers?.bg).toBeInstanceOf(Float32Array)
+    expect(frames[0].buffers?.bg).toBeInstanceOf(Uint16Array)
     expect(frames[0].buffers?.fg).toBeUndefined()
     expect(frames[0].buffers?.attributes).toBeUndefined()
 
@@ -349,8 +348,8 @@ describe("TestRecorder", () => {
     const frames = recorderWithAll.recordedFrames
     expect(frames.length).toBe(1)
     expect(frames[0].buffers).toBeDefined()
-    expect(frames[0].buffers?.fg).toBeInstanceOf(Float32Array)
-    expect(frames[0].buffers?.bg).toBeInstanceOf(Float32Array)
+    expect(frames[0].buffers?.fg).toBeInstanceOf(Uint16Array)
+    expect(frames[0].buffers?.bg).toBeInstanceOf(Uint16Array)
     expect(frames[0].buffers?.attributes).toBeInstanceOf(Uint8Array)
 
     recorderWithAll.stop()
